refactor: extract ExpressionDecoder, StatementDecoder, FieldDecoder from Decoder (#480)
Decompose Decoder.java (1258 lines) into focused delegate classes:
- ExpressionDecoder (366 lines): value expressions, comparisons, arithmetic, returns
- StatementDecoder (449 lines): method/constructor bodies, control flow, assignments
- FieldDecoder (194 lines): field declarations, initializers

Decoder.java is now a thin coordinator (343 lines, down from 1258) that
delegates to the three new package-private classes. All delegates call back
to Decoder for shared services (type resolution, scope finders, primitiveLiteral).

No public API changes to TweedleEncoderDecoder. All existing tests pass:
- TweedleEncoderDecoderTest (core/ast)
- DecoderDelegateDecompositionCharacterizationTest (core/ast)
- SilverThreadTweedleDecoderRoundTripTest (core/story-api-migration)
